requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
huggingface_hub>=0.16.0
streamlit>=1.25.0
Pillow>=9.5.0
numpy>=1.24.0
matplotlib>=3.7.0
nltk>=3.8.1
requests>=2.31.0
pandas>=2.0.0
accelerate>=0.21.0
bitsandbytes>=0.40.0
sentencepiece>=0.1.99 